requirements.txt
streamlit==1.28.1
pandas==2.1.3
plotly==5.18.0
scikit-learn==1.3.2
google-generativeai==0.3.0
python-dotenv==1.0.0
openpyxl==3.1.2
numpy==1.24.3
